requirements.txt
Python==3.8
Django==4.2.9
python-dotenv==1.0.1
django-filter==23.5
django-allauth==0.54.0
django-apscheduler==0.6.2
celery==5.3.6
redis==5.0.3
whitenoise==6.6.0
gettext==5.0
django-modeltranslation==0.19.3